feat(): display tags on read --short
This commits adds the ability to finally return the tags contain in
title and body of the entries when reading entries.
When the user do a reading of the entries with the --short flag active,
the output will display the tags contain within each displayed entry.

The commit also fixes a bug that made writing titles with puntuation
marks (.!?) don't write the mark when saving the entry.