feat(): add style format with LGG_STYLE env
This commit changes the undocumented flag --short for --style, an enum
that allows the user to set the style on which the entries are printed
to the terminal.
It allows setting a default through the env variable LGG_STYLE.
This commit also refactors the CLI struct, moving it to its own module
for better maintanability.